Refactor some Adpcm code
