+++ src/VGAudio/Utilities/Endianness.cs
@@ -1,0 +1,8 @@
+namespace VGAudio.Utilities
+{
+    public enum Endianness
+    {
+        BigEndian,
+        LittleEndian
+    }
+}
