+++ src/VGAudio/Utilities/Helpers.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.IO;
+﻿using System.IO;
 using System.Runtime.CompilerServices;
 
 namespace VGAudio.Utilities
@@ -30,43 +29,9 @@ namespace VGAudio.Utilities
             return value + multiple - value % multiple;
         }
 
-        internal static void CheckStream(Stream stream, int minLength)
-        {
-            if (!stream.CanSeek)
-            {
-                throw new NotSupportedException("A seekable stream is required");
-            }
-
-            if (stream.Length < minLength)
-            {
-                throw new InvalidDataException($"File is only {stream.Length} bytes long");
-            }
-        }
-
-        internal static bool ArraysEqual<T>(T[] a1, T[] a2)
-        {
-            if (a1 == null || a2 == null) return false;
-            if (a1 == a2) return true;
-            if (a1.Length != a2.Length) return false;
-
-            for (int i = 0; i < a1.Length; i++)
-            {
-                if (!a1[i].Equals(a2[i]))
-                {
-                    return false;
-                }
-            }
-            return true;
-        }
 
         public static bool LoopPointsAreAligned(int loopStart, int alignmentMultiple)
             => !(alignmentMultiple != 0 && loopStart % alignmentMultiple != 0);
-
-        public enum Endianness
-        {
-            BigEndian,
-            LittleEndian
-        }
 
         public static BinaryReader GetBinaryReader(Stream stream, Endianness endianness) =>
             endianness == Endianness.LittleEndian
